Refactor web resources
- Change API to send array of tuples instead of object
