+++ loft/web/static/src/base.css
@@ -1,47 +1,8 @@
 
-:root {
-    --font-base: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-    --font-header: 'Courier New', Courier, monospace;
-
-    /* color scheme */
-    --color-bg: white;
-    --color-base: var(--color-blue-1);
-    --color-primary-fg: var(--color-base);
-    --color-primary-bg: var(--color-green-3);
-    --color-secondary-fg: var(--color-base);
-    --color-secondary-bg: var(--color-blue-3);
-    --color-tertiary-fg: var(--color-base);
-    --color-tertiary-bg: var(--color-gray-3);
-
-    --color-accent-fg: var(--color-base);
-    --color-accent-bg: var(--color-green-45);
-    --color-accent-active: var(--color-purple-3);
-
-    --color-secondary-accent-fg: var(--color-base);
-    --color-secondary-accent-bg: var(--color-blue-45);
-    --color-secondary-accent-active: var(--color-purple-3);
-
-    --color-success-fg: var(--color-green-3);
-    --color-success-bg: var(--color-base);
-    --color-error-fg: var(--color-red-3);
-    --color-error-bg: var(--color-base);
-
-    --color-blue-1: #e9f4fa;
-    --color-blue-3: #6792aa;    /* 3.0 contrast ratio */
-    --color-blue-45: #557384;   /* 4.5 contrast ratio */
-
-    --color-red-3: #e1665f;     /* 3.0 contrast ratio */
-    --color-red-45: #c14541;    /* 4.5 contrast ratio */
-
-    --color-green-3: #309c95;   /* 3.0 contrast ratio */
-    --color-green-45: #367a75;  /* 4.5 contrast ratio */
-
-    --color-gray-3: #8d8c8c;    /* 3.0 contrast ratio */
-    --color-gray-45: #6f6f6f;   /* 4.5 contrast ratio */
-
-    --color-purple-3: #b374ca;  /* 3.0 contrast ratio */
-    --color-purple-45: #a83ccb; /* 4.5 contrast ratio */
-}
+@import './css/bubbles.css';
+@import './css/groups.css';
+@import './css/messages.css';
+@import './css/selection.css';
 
 html {
     font-family: var(--font-base);
@@ -62,54 +23,6 @@ h1 {
     background-color: var(--color-error-bg);
 }
 
-.button, .bubble, .flash {
-    width: 100%;
-    margin: .25em;
-    padding: .5em 1em;
-    display: block;
-
-    box-sizing: border-box;
-}
-
-.button {
-    font-size: 100%;
-    text-align: center;
-    text-decoration: none;
-
-    cursor: pointer;
-
-    color: var(--color-primary-fg);
-    background-color: var(--color-primary-bg);
-    border: none;
-    border-radius: 4px;
-
-    transition: color 0.2s ease;
-    transition: background-color 0.2s ease;
-}
-
-.button-secondary {
-    color: var(--color-secondary-fg);
-    background-color: var(--color-secondary-bg);
-}
-
-.button:hover, .button:focus {
-    color: var(--color-accent-fg);
-    background-color: var(--color-accent-bg);
-}
-
-.button:active {
-    color: var(--color-base);
-    background-color: var(--color-accent-active);
-}
-
-.button-secondary:hover, .button-secondary:focus {
-    color: var(--color-secondary-accent-fg);
-    background-color: var(--color-secondary-accent-bg);
-}
-
-.button-secondary:active {
-    background-color: var(--color-secondary-accent-active);
-}
 
 .content {
     margin: auto;
@@ -117,58 +30,4 @@ h1 {
     top: 50%;
     left: 50%;
     transform: translate(-50%, -50%);
-}
-
-.error, .success {
-    border-radius: 4px;
-}
-
-.error {
-    color: var(--color-error-fg);
-    background-color: var(--color-error-bg);
-}
-
-.success {
-    color: var(--color-success-fg);
-    background-color: var(--color-success-bg);
-}
-
-.success {
-    color: var(--color-success-fg);
-    background-color: var(--color-success-bg);
-}
-
-.messages div {
-    text-align: center;
-
-    font-style: italic;
-}
-
-.group {
-    width: 60vw;
-    margin: .5rem;
-    padding: 1.5rem;
-
-    border-radius: 1rem;
-    background-color: var(--color-base);
-
-    box-sizing: border-box;
-}
-
-/* medium screens and larger */
-@media screen and (min-width: 36em) {
-    .group {
-        width: 24em;
-    }
-}
-
-.selected-file {
-    font-size: 100%;
-    text-align: center;
-    text-decoration: none;
-
-    color: var(--color-tertiary-fg);
-    background-color: var(--color-tertiary-bg);
-    border: none;
-    border-radius: 4px;
 }
