+++ loft/web/static/src/css/groups.css
@@ -1,0 +1,18 @@
+
+.group {
+    width: 60vw;
+    margin: .5rem;
+    padding: 1.5rem;
+
+    border-radius: 1rem;
+    background-color: var(--color-base);
+
+    box-sizing: border-box;
+}
+
+/* medium screens and larger */
+@media screen and (min-width: 36em) {
+    .group {
+        width: 24em;
+    }
+}
